requirements.txt
google-cloud-dialogflow==2.13.0
python-telegram-bot==13.11
vk-api==11.9.8
python-dotenv==0.19.2